Use committer date instead of author date
Author date represents the date the commit was first created. In case a
branch hash squashed commits, it would normally have the date of oldest
commit as author date but the timestamp of the squash as the committer
date. Thus committer date is a better measure of determining branch age.